requirements.txt
python-telegram-bot==20.7
aiohttp==3.9.0
pydantic==2.5.0
python-dotenv==1.0.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.5
pytest-cov==4.1.0
pytest-mock==3.14.0

# Type checking
mypy==1.8.0

